+++ src/features/profile/components/clubs/index.ts
@@ -1,4 +1,1 @@
-export { default as CreatedClubList } from './CreatedClubsList';
-export { default as JoinedClubList } from './JoinedClubList';
-export { default as MyReviewList } from './MyReviewList';
 export { default as WriteReviewModal } from './WriteReviewModal';
