+++ src/features/profile/container/index.ts
@@ -2,3 +2,6 @@ export { default as ClubContents } from './ClubContents';
 export { default as MainContent } from './MainContent';
 export { default as ProfileHeader } from './ProfileHeader';
 export { default as ProfilePage } from './ProfilePage';
+export { default as MyReviewList } from './MyReviewList';
+export { default as JoinedClubList } from './JoinedClubList';
+export { default as CreatedClubList } from './CreatedClubsList';
